+++ src/apps/triangles/mod.rs
@@ -1,2 +1,4 @@
 pub mod triangles;
+pub mod geometric_utils;
+
 pub use triangles::TrianglesApp;
